feat(alloy): Migrate pool lens to use Alloy (#2)
* Update build process and refactor project configuration

This commit makes significant changes to the build process and project configuration. The changes include updating the .gitignore file to include Cargo.lock, adding the INFURA_API_KEY environmental variable to Rust workflow, and modifying commands in the Rust and Github workflows. Refactoring also involved changing the project's metadata in Cargo.toml and altering various settings in rustfmt.toml and package.json. Lastly, the Github workflow for npm_publish was removed.

* feat(alloy): Migrate pool lens to use Alloy

This commit migrates the pool lens functions to use Alloy for making calls and decoding results. All calls to the `Middleware` trait and its `client` have been replaced with the Alloy `Provider` trait and its `provider`. The functions now return an `anyhow::Result` instead of a `ContractError`, providing a more generic error handling approach. The diff also includes commenting out `position_lens` and `storage_lens`. Further testing and validation are required to ensure correct information retrieval and error management post-migration.

* Remove formatting check from GitHub Actions

The commit removes the formatting check step from the GitHub Actions workflow for Rust. This step was unnecessary and was causing redundancy in the workflow. The focus is now solely on linting checks, making the process more streamlined and efficient.

* fix ci

* fix test

* fix fmt