feat(alloy): update alloy to 0.3 (#6)
* Add alias types and upgrade dependencies

Replaced primitive integer types with alias types (U24 and I24) for better clarity and consistency. Enabled `clippy::redundant_clone` lint. Updated `alloy` dependency and incremented package version to 0.2.0.

* Add detailed documentation to lens modules

Enhance the position, pool, and storage lens modules in the code with comprehensive documentation. This includes descriptions of functionalities and explanations of arguments and return values for functions. Additionally, update the README and Cargo.toml to reflect the library's purpose and features accurately.

* Fix test

* Remove PCSV3-specific tests

Deleted tests related to PancakeSwap V3 for BSC in both TypeScript and Solidity test files. This includes the removal of the PCSV3-specific PoolLens and StorageLens tests as well as updates to the gas-snapshot file to reflect these deletions.