feat: add `no_std` feature and update dependencies (#8)
* Add `no_std` support and update dependencies

Introduces `no_std` compatibility to the library and includes necessary updates to dependencies in `Cargo.toml`. This change enables the crate to be used in environments without the standard library, expanding its applicability.

* Remove "transports" feature from alloy dependency

The "transports" feature in the alloy dependency is unnecessary and has been removed. This change helps streamline the dependencies and potentially reduces the binary size. All current functionality remains intact without this feature.

* Simplify `BLOCK_NUMBER` usage

Refactor code to remove unnecessary dereferencing of BLOCK_NUMBER across multiple modules. This enhances readability and adheres to Rust best practices by directly passing the constant reference.

* Add note on `no_std` support in README

This update explains that the library, by default, does not depend on the standard library. It also mentions that the `std` feature can be enabled using the `std` feature flag.